+++ src/bitmessagemain.py
@@ -22,10 +22,13 @@ depends.check_dependencies()
 
 import ctypes
 import getopt
+import multiprocessing
 # Used to capture a Ctrl-C keypress so that Bitmessage can shutdown gracefully.
 import signal
 import socket
+import threading
 import time
+import traceback
 from struct import pack
 
 from helper_startup import (
@@ -38,7 +41,7 @@ import shared
 import knownnodes
 import state
 import shutdown
-import threading
+from debug import logger
 
 # Classes
 from class_sqlThread import sqlThread
@@ -61,8 +64,8 @@ from network.downloadthread import DownloadThread
 from network.uploadthread import UploadThread
 
 # Helper Functions
-import helper_generic
 import helper_threading
+
 
 def connectToStream(streamNumber):
     state.streamsInWhichIAmParticipating.append(streamNumber)
@@ -148,6 +151,43 @@ def _fixSocket():
         socket.IPPROTO_IPV6 = 41
     if not hasattr(socket, 'IPV6_V6ONLY'):
         socket.IPV6_V6ONLY = 27
+
+
+def allThreadTraceback(frame):
+    id2name = dict([(th.ident, th.name) for th in threading.enumerate()])
+    code = []
+    for threadId, stack in sys._current_frames().items():
+        code.append(
+            '\n# Thread: %s(%d)' % (id2name.get(threadId, ''), threadId))
+        for filename, lineno, name, line in traceback.extract_stack(stack):
+            code.append(
+                'File: "%s", line %d, in %s' % (filename, lineno, name))
+            if line:
+                code.append('  %s' % (line.strip()))
+    print('\n'.join(code))
+
+
+def signal_handler(signal, frame):
+    process = multiprocessing.current_process()
+    logger.error(
+        'Got signal %i in %s/%s',
+        signal, process.name, threading.current_thread().name
+    )
+    if process.name == "RegExParser":
+        # on Windows this isn't triggered, but it's fine,
+        # it has its own process termination thing
+        raise SystemExit
+    if "PoolWorker" in process.name:
+        raise SystemExit
+    if threading.current_thread().name not in ("PyBitmessage", "MainThread"):
+        return
+    logger.error("Got signal %i", signal)
+    if shared.thisapp.daemon or not state.enableGUI:  # FIXME redundant?
+        shutdown.doCleanShutdown()
+    else:
+        allThreadTraceback(frame)
+        print('Unfortunately you cannot use Ctrl+C when running the UI'
+              ' because the UI captures the signal.')
 
 
 # This is a list of current connections (the thread pointers at least)
@@ -437,8 +477,8 @@ class Main:
             os.kill(grandfatherPid, signal.SIGTERM)
 
     def setSignalHandler(self):
-        signal.signal(signal.SIGINT, helper_generic.signal_handler)
-        signal.signal(signal.SIGTERM, helper_generic.signal_handler)
+        signal.signal(signal.SIGINT, signal_handler)
+        signal.signal(signal.SIGTERM, signal_handler)
         # signal.signal(signal.SIGINT, signal.SIG_DFL)
 
     def usage(self):
